// frontend-web/src/utils/transport-action-enum.ts
export enum TransportActionEnum {
    INIT_USER_DATA = "INIT_USER_DATA",
    CONNECT = "CONNECT",
    FETCH_GROUP_MESSAGES = "FETCH_GROUP_MESSAGES",
    SEND_GROUP_MESSAGE = "SEND_GROUP_MESSAGE",
    NOTIFICATION_MESSAGE = "NOTIFICATION_MESSAGE",
    GRANT_USER_ADMIN = "GRANT_USER_ADMIN",
    ADD_CHAT_HISTORY = "ADD_CHAT_HISTORY",
    MARK_MESSAGE_AS_SEEN = "MARK_MESSAGE_AS_SEEN",

    INIT_VIDEO_CALL = "INIT_VIDEO_CALL",
    ACCEPT_VIDEO_CALL = "ACCEPT_VIDEO_CALL",
    START_VIDEO_CALL = "START_VIDEO_CALL",

    RTC_USER_ANSWER = "RTC_USER_ANSWER",
    RTC_USER_OFFER = "RTC_USER_OFFER",
    RTC_NEW_ICE_CANDIDATE = "RTC_NEW_ICE_CANDIDATE",
    RTC_READY_STATE = "RTC_READY_STATE"
}